Address PR #299 review: make classifier tags authoritative; split test file
- reportPresetError: flip the merge so NetworkErrorClassifier.errorTags win over
  caller-supplied extraTags on key conflict (Greptile P2 — keeps Sentry
  diagnostic tags authoritative).
- Move the two new preset error-reporting tests into
  StationListPresetErrorReportingTests.swift so StationListPresetTests.swift
  stays under the 1000-line SwiftLint file_length limit; expose the shared
  signedInAuth / makePresetVisibleItem helpers for reuse.
