+++ PlayolaRadio/Views/Pages/StationListPage/StationListPresetTests.swift
@@ -221,60 +221,6 @@ struct StationListPresetTests {
   }
 
   @Test
-  func testStarTappedAddNetworkErrorIsNotReportedToSentry() async {
-    @Shared(.auth) var auth = signedInAuth()
-    @Shared(.presets) var presets: IdentifiedArrayOf<Preset> = []
-    @Shared(.pendingPresetStationIds) var pending: Set<String> = []
-
-    let item = makePresetVisibleItem()
-    let reportCount = LockIsolated(0)
-
-    let model = withDependencies {
-      $0.api.createPreset = { _, _, _ in
-        throw NSError(domain: NSURLErrorDomain, code: NSURLErrorSecureConnectionFailed)
-      }
-      $0.analytics.track = { _ in }
-      $0.errorReporting.reportError = { _, _ in reportCount.setValue(reportCount.value + 1) }
-    } operation: {
-      StationListModel()
-    }
-
-    await model.starTapped(for: item)
-
-    #expect(reportCount.value == 0)
-    #expect(pending.isEmpty)
-    #expect(presets.isEmpty)
-    #expect(model.presentedAlert == .errorSavingPreset(nil))
-  }
-
-  @Test
-  func testStarTappedAddTagsReportedErrorWithDomainAndCode() async {
-    @Shared(.auth) var auth = signedInAuth()
-    @Shared(.presets) var presets: IdentifiedArrayOf<Preset> = []
-    @Shared(.pendingPresetStationIds) var pending: Set<String> = []
-
-    let item = makePresetVisibleItem()
-    let reportedTags = LockIsolated<[String: String]>([:])
-
-    let model = withDependencies {
-      $0.api.createPreset = { _, _, _ in
-        throw NSError(domain: "TestDomain", code: 42)
-      }
-      $0.analytics.track = { _ in }
-      $0.errorReporting.reportError = { _, tags in reportedTags.setValue(tags) }
-    } operation: {
-      StationListModel()
-    }
-
-    await model.starTapped(for: item)
-
-    #expect(reportedTags.value["endpoint"] == "POST /v1/presets")
-    #expect(reportedTags.value["error_domain"] == "TestDomain")
-    #expect(reportedTags.value["error_code"] == "42")
-    #expect(reportedTags.value["station_id"] == item.anyStation.id)
-  }
-
-  @Test
   func testStarTappedIgnoredWhilePendingAdd() async {
     @Shared(.auth) var auth = signedInAuth()
     @Shared(.presets) var presets: IdentifiedArrayOf<Preset> = []
@@ -1012,7 +958,7 @@ struct StationListPresetTests {
   }
 }
 
-private func signedInAuth() -> Auth {
+func signedInAuth() -> Auth {
   Auth(
     currentUser: LoggedInUser(
       id: "u1", firstName: "B", lastName: nil, email: "b@x.com",
@@ -1038,7 +984,7 @@ private func makePresetTestList(with items: [APIStationItem], date: Date = Date(
     hidden: false, sortOrder: 0, createdAt: date, updatedAt: date, items: items)
 }
 
-private func makePresetVisibleItem(date: Date = Date()) -> APIStationItem {
+func makePresetVisibleItem(date: Date = Date()) -> APIStationItem {
   APIStationItem(
     sortOrder: 0,
     visibility: .visible,
